Splitting views into different files
