+++ popular_proposal/views/proposal_views.py
@@ -1,46 +1,54 @@
 # coding=utf-8
-from django.views.generic.edit import FormView, UpdateView
-from popular_proposal.forms import (ProposalForm,
-                                    SubscriptionForm,
-                                    get_form_list,
-                                    AreaForm,
-                                    UpdateProposalForm,
-                                    ProposalFilterForm,
-                                    )
-from django.core.urlresolvers import reverse
-from django.shortcuts import get_object_or_404
-from django.utils.decorators import method_decorator
+
+from backend_candidate.models import Candidacy
+
+from constance import config
+
+from django.contrib import messages
+
 from django.contrib.auth.decorators import login_required
-from django.views.generic.base import TemplateView
-from django.views.generic.detail import DetailView
-from popular_proposal.models import (PopularProposal,
-                                     ProposalTemporaryData,
-                                     Commitment,
-                                     ProposalLike)
-from django.shortcuts import render_to_response
-from formtools.wizard.views import SessionWizardView
-from collections import OrderedDict
+
+from django.core.urlresolvers import reverse
+
+from django.http import HttpResponseNotFound, JsonResponse
+
+from django.shortcuts import get_object_or_404
+
+from django.template.response import TemplateResponse
+
+from django.utils.decorators import method_decorator
+
+from django.utils.translation import ugettext as _
+
 from django.views.generic import View
-from django.http import JsonResponse, HttpResponseNotFound
-from django_filters.views import FilterView
+
+from django.views.generic.detail import DetailView
+
+from django.views.generic.edit import FormView, UpdateView
+
 from django.views.generic.list import ListView
-from popular_proposal.forms import ProposalAreaFilterForm
+
+from django_filters.views import FilterView
+
+from elections.models import Area, Candidate
+
 from popular_proposal.filters import ProposalAreaFilter
-from votainteligente.view_mixins import EmbeddedViewBase
-from votainteligente.send_mails import send_mails_to_staff
+
 from popular_proposal.forms import (CandidateCommitmentForm,
                                     CandidateNotCommitingForm,
+                                    ProposalAreaFilterForm,
+                                    ProposalFilterForm,
+                                    ProposalForm,
+                                    UpdateProposalForm,
+                                    SubscriptionForm,
                                     )
-from elections.models import Candidate, Area
-from backend_candidate.models import (Candidacy,
-                                      is_candidate,
-                                      )
-from django.conf import settings
-from django.contrib import messages
-from django.utils.translation import ugettext as _
-from django.template.response import TemplateResponse
-from django.shortcuts import render
-from constance import config
+
+from popular_proposal.models import (Commitment,
+                                     PopularProposal,
+                                     ProposalLike,
+                                     ProposalTemporaryData,)
+
+from votainteligente.view_mixins import EmbeddedViewBase
 
 
 class ProposalCreationView(FormView):
@@ -68,7 +76,8 @@ class ProposalCreationView(FormView):
         return super(ProposalCreationView, self).form_valid(form)
 
     def get_success_url(self):
-        return reverse('popular_proposals:thanks', kwargs={'pk': self.proposal.id})
+        return reverse('popular_proposals:thanks',
+                       kwargs={'pk': self.proposal.id})
 
 
 class ThanksForProposingView(DetailView):
@@ -88,7 +97,8 @@ class SubscriptionView(FormView):
 
     @method_decorator(login_required)
     def dispatch(self, *args, **kwargs):
-        self.proposal = get_object_or_404(PopularProposal, id=self.kwargs['pk'])
+        self.proposal = get_object_or_404(PopularProposal,
+                                          id=self.kwargs['pk'])
         if self.request.method == 'GET':
             self.next_url = self.request.GET.get('next', None)
         elif self.request.method == 'POST':
@@ -112,7 +122,9 @@ class SubscriptionView(FormView):
         like = form.subscribe()
         context = self.get_context_data()
         context['like'] = like
-        return TemplateResponse(self.request, 'popular_proposal/subscribing_result.html', context)
+        return TemplateResponse(self.request,
+                                'popular_proposal/subscribing_result.html',
+                                context)
 
 
 class HomeView(EmbeddedViewBase, FilterView):
@@ -138,136 +150,6 @@ class PopularProposalDetailView(EmbeddedViewBase, DetailView):
     model = PopularProposal
     template_name = 'popular_proposal/detail.html'
     context_object_name = 'popular_proposal'
-
-wizard_form_list = get_form_list()
-
-
-class ProposalWizardBase(SessionWizardView):
-    form_list = wizard_form_list
-    template_name = 'popular_proposal/wizard/form_step.html'
-
-    def get_template_names(self):
-        form = self.get_form(step=self.steps.current)
-        template_name = getattr(form, 'template', self.template_name)
-        return template_name
-
-    def get_previous_forms(self):
-        return []
-
-    def get_form_list(self):
-        form_list = OrderedDict()
-        previous_forms = self.get_previous_forms()
-        my_list = previous_forms + get_form_list(user=self.request.user)
-        counter = 0
-        for form_class in my_list:
-            form_list[str(counter)] = form_class
-            counter += 1
-        self.form_list = form_list
-        return form_list
-
-    def determine_area(self, data):
-        if 'area' in data.keys():
-            return data['area']
-        elif hasattr(self, 'area'):
-            return self.area
-        else:
-            return Area.objects.get(id=config.DEFAULT_AREA)
-
-    def done(self, form_list, **kwargs):
-        data = {}
-        [data.update(form.cleaned_data) for form in form_list]
-        kwargs = {
-            'proposer': self.request.user,
-            'data': data
-        }
-        
-        kwargs['area'] = self.determine_area(data)
-        temporary_data = ProposalTemporaryData.objects.create(**kwargs)
-        if not settings.MODERATION_ENABLED:
-            temporary_data.create_proposal()
-        else:
-            temporary_data.notify_new()
-        context = self.get_context_data(form=None)
-        context.update({'popular_proposal': temporary_data,
-                        'area': kwargs['area']
-                        })
-        send_mails_to_staff({'temporary_data': temporary_data}, 'notify_staff_new_proposal')
-        return render_to_response('popular_proposal/wizard/done.html',
-                                  context)
-
-    def get_context_data(self, form, **kwargs):
-        context = super(ProposalWizardBase, self).get_context_data(form, **kwargs)
-        data = self.get_all_cleaned_data()
-        if data:
-            context['area'] = self.determine_area(data)
-        context['preview_data'] = data
-        return context
-
-    def get_form_kwargs(self, step=None):
-        kwargs = super(ProposalWizardBase, self).get_form_kwargs(step)
-        kwargs['is_staff'] = self.request.user.is_staff
-        return kwargs
-
-    @method_decorator(login_required)
-    def dispatch(self, request, *args, **kwargs):
-        if is_candidate(request.user):
-            return HttpResponseNotFound()
-        if config.PROPOSALS_ENABLED:
-            return super(ProposalWizardBase, self).dispatch(request, *args, **kwargs)
-        else:
-            return HttpResponseNotFound()
-
-
-class ProposalWizard(ProposalWizardBase):
-    '''
-    Esta es la clase del wizard a la que se llega por hacer
-    /propuestas/crear/villarrica
-    '''
-    @method_decorator(login_required)
-    def dispatch(self, request, *args, **kwargs):
-        self.area = get_object_or_404(Area, id=self.kwargs['slug'])
-        return super(ProposalWizard, self).dispatch(request, *args, **kwargs)
-
-
-class ProposalWizardFull(ProposalWizardBase):
-    '''
-    Esta es la clase del wizard a la que se llega por hacer
-    /propuestas/create_full_wizard
-    Acá lo primero que te preguntamos es para qué comuna quieres crear la propuesta
-    '''
-    form_list = [AreaForm, ] + wizard_form_list
-
-    def get_previous_forms(self):
-        return [AreaForm, ]
-
-
-class ProposalWizardFullWithoutArea(ProposalWizardBase):
-    '''
-    Esta es la clase del wizard a la que se llega por hacer
-    /propuestas/crear
-    Acá no te preguntamos por el area por que sabemos que es la que viene por defecto.
-    '''
-    form_list = wizard_form_list
-
-    def get_previous_forms(self):
-        return []
-
-class PopularProposalUpdateView(UpdateView):
-    form_class = UpdateProposalForm
-    template_name = 'popular_proposal/update.html'
-    model = PopularProposal
-    context_object_name = 'popular_proposal'
-
-    @method_decorator(login_required)
-    def dispatch(self, request, *args, **kwargs):
-        return super(PopularProposalUpdateView, self).dispatch(request,
-                                                               *args,
-                                                               **kwargs)
-
-    def get_queryset(self):
-        qs = super(PopularProposalUpdateView, self).get_queryset()
-        qs = qs.filter(proposer=self.request.user)
-        return qs
 
 
 class UnlikeProposalView(View):
@@ -319,13 +201,17 @@ class CommitView(FormView):
     def dispatch(self, *args, **kwargs):
         if not config.PROPOSALS_ENABLED:
             return HttpResponseNotFound()
-        self.proposal = get_object_or_404(PopularProposal, id=self.kwargs['proposal_pk'])
-        self.candidate = get_object_or_404(Candidate, id=self.kwargs['candidate_pk'])
+        self.proposal = get_object_or_404(PopularProposal,
+                                          id=self.kwargs['proposal_pk'])
+        self.candidate = get_object_or_404(Candidate,
+                                           id=self.kwargs['candidate_pk'])
         previous_commitment_exists = Commitment.objects.filter(proposal=self.proposal,
                                                                candidate=self.candidate).exists()
         if previous_commitment_exists:
             return HttpResponseNotFound()
-        get_object_or_404(Candidacy, candidate=self.candidate, user=self.request.user)
+        get_object_or_404(Candidacy,
+                          candidate=self.candidate,
+                          user=self.request.user)
         # The following can be refactored
         areas = []
         for election in self.candidate.elections.all():
@@ -388,3 +274,21 @@ class CommitmentDetailView(DetailView):
 
     def get_object(self, queryset=None):
         return get_object_or_404(self.model, candidate=self.candidate, proposal=self.proposal)
+
+
+class PopularProposalUpdateView(UpdateView):
+    form_class = UpdateProposalForm
+    template_name = 'popular_proposal/update.html'
+    model = PopularProposal
+    context_object_name = 'popular_proposal'
+
+    @method_decorator(login_required)
+    def dispatch(self, request, *args, **kwargs):
+        return super(PopularProposalUpdateView, self).dispatch(request,
+                                                               *args,
+                                                               **kwargs)
+
+    def get_queryset(self):
+        qs = super(PopularProposalUpdateView, self).get_queryset()
+        qs = qs.filter(proposer=self.request.user)
+        return qs
